+++ src/main/java/kitchenpos/table/domain/repository/OrderTableRepository.java
@@ -1,6 +1,7 @@
-package kitchenpos.table.domain;
+package kitchenpos.table.domain.repository;
 
 import java.util.List;
+import kitchenpos.table.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
