+++ src/main/java/kitchenpos/table/service/TableService.java
@@ -2,7 +2,7 @@ package kitchenpos.table.service;
 
 import java.util.List;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.table.domain.repository.OrderTableRepository;
 import kitchenpos.table.service.dto.TableRequest;
 import kitchenpos.table.service.dto.TableResponse;
 import org.springframework.stereotype.Service;
