+++ src/main/java/kitchenpos/table/service/dto/TableIdRequest.java
@@ -1,0 +1,14 @@
+package kitchenpos.table.service.dto;
+
+public class TableIdRequest {
+
+    private long id;
+
+    public TableIdRequest(final long id) {
+        this.id = id;
+    }
+
+    public long getId() {
+        return id;
+    }
+}
