--- src/main/java/kitchenpos/tablegroup/service/dto/OrderTableIdRequest.java
@@ -1,14 +1,0 @@
-package kitchenpos.tablegroup.service.dto;
-
-public class OrderTableIdRequest {
-
-    private long id;
-
-    public OrderTableIdRequest(final long id) {
-        this.id = id;
-    }
-
-    public long getId() {
-        return id;
-    }
-}
