--- src/main/java/kitchenpos/tablegroup/service/dto/TableGroupRequest.java
@@ -1,19 +1,0 @@
-package kitchenpos.tablegroup.service.dto;
-
-import java.util.List;
-
-public class TableGroupRequest {
-
-    private List<OrderTableIdRequest> orderTables;
-
-    public TableGroupRequest() {
-    }
-
-    public TableGroupRequest(final List<OrderTableIdRequest> orderTables) {
-        this.orderTables = orderTables;
-    }
-
-    public List<OrderTableIdRequest> getOrderTables() {
-        return orderTables;
-    }
-}
