+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -6,12 +6,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import kitchenpos.order.service.dto.OrderStatusRequest;
-import kitchenpos.tablegroup.service.dto.TableGroupRequest;
+import kitchenpos.table.service.dto.TableGroupRequest;
 import kitchenpos.menugroup.service.dto.MenuGroupResponse;
 import kitchenpos.menu.service.dto.MenuResponse;
 import kitchenpos.order.service.dto.OrderResponse;
 import kitchenpos.product.service.dto.ProductResponse;
-import kitchenpos.tablegroup.service.dto.TableGroupResponse;
+import kitchenpos.table.service.dto.TableGroupResponse;
 import kitchenpos.table.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
@@ -25,7 +25,7 @@ import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
 import kitchenpos.supports.TableGroupFixture;
 import kitchenpos.table.service.TableService;
-import kitchenpos.tablegroup.service.TableGroupService;
+import kitchenpos.table.service.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
