+++ src/test/java/kitchenpos/supports/TableGroupFixture.java
@@ -3,22 +3,22 @@ package kitchenpos.supports;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.tablegroup.service.dto.OrderTableIdRequest;
-import kitchenpos.tablegroup.service.dto.TableGroupRequest;
+import kitchenpos.table.service.dto.TableIdRequest;
+import kitchenpos.table.service.dto.TableGroupRequest;
 import kitchenpos.table.service.dto.TableResponse;
 
 public class TableGroupFixture {
 
     public static TableGroupRequest from(final TableResponse... orderTables) {
-        final List<OrderTableIdRequest> tableIds = Arrays.stream(orderTables)
-                .map(each -> new OrderTableIdRequest(each.getId()))
+        final List<TableIdRequest> tableIds = Arrays.stream(orderTables)
+                .map(each -> new TableIdRequest(each.getId()))
                 .collect(Collectors.toList());
         return new TableGroupRequest(tableIds);
     }
 
     public static TableGroupRequest fromIds(final Long... tableIds) {
-        final List<OrderTableIdRequest> tableIdRequests = Arrays.stream(tableIds)
-                .map(OrderTableIdRequest::new)
+        final List<TableIdRequest> tableIdRequests = Arrays.stream(tableIds)
+                .map(TableIdRequest::new)
                 .collect(Collectors.toList());
         return new TableGroupRequest(tableIdRequests);
     }
